+++ src/main/java/com/enfernuz/quik/lua/rpc/serde/protobuf/BXorArgsPbSerializer.java
@@ -5,7 +5,7 @@ import com.enfernuz.quik.lua.rpc.serde.Serializer;
 import org.jetbrains.annotations.NotNull;
 import qlua.rpc.bit.Bxor;
 
-import java.util.stream.IntStream;
+import java.util.stream.LongStream;
 
 import static java.util.stream.Collectors.toList;
 
@@ -28,10 +28,10 @@ enum BXorArgsPbSerializer implements Serializer<BXor.Args>, ToPbConverter<BXor.A
                     .setX1( args.getX1() )
                     .setX2( args.getX2() );
 
-        final int[] xi = args.getXi();
+        final long[] xi = args.getXi();
         if (xi != null) {
             result.addAllXi(
-                    IntStream.of(xi)
+                    LongStream.of(xi)
                             .boxed()
                             .collect( toList() )
             );
