+++ src/test/java/com/enfernuz/quik/lua/rpc/serde/json/BitBOrJsonSerdeTest.java
@@ -15,7 +15,7 @@ public class BitBOrJsonSerdeTest {
             return BOr.Args.builder()
                     .x1(1)
                     .x2(2)
-                    .xi(new int[] {3, 5, 7, 9})
+                    .xi(new long[] {3L, 5L, 7L, 9L})
                     .build();
         }
 
@@ -31,8 +31,8 @@ public class BitBOrJsonSerdeTest {
         public BOr.Args getArgsObject() {
 
             return BOr.Args.builder()
-                    .x1(1)
-                    .x2(2)
+                    .x1(1L)
+                    .x2(2L)
                     .build();
         }
